[warning-panel] Addressed PR comments
