+++ packages/perseus-editor/src/item-editor.tsx
@@ -22,7 +22,6 @@ type Props = {
     apiOptions?: APIOptions;
     deviceType?: DeviceType;
     widgetIsOpen?: boolean;
-    widgetAccessibility: boolean;
     gradeMessage?: string;
     imageUploader?: ImageUploader;
     wasAnswered?: boolean;
@@ -99,7 +98,7 @@ class ItemEditor extends React.Component<Props> {
             <div className="perseus-editor-table">
                 <div className="perseus-editor-row perseus-question-container">
                     <div className="perseus-editor-left-cell">
-                        <IssuesPanel />
+                        <IssuesPanel violations={[]} incompletes={[]}/>
                         <div className="pod-title">Question</div>
                         <Editor
                             ref={this.questionEditor}
